4. Шаблонизируй это: вынес общие функции в utils.js, добавил функционал для кнопкий "load more films"
