+++ src/mock/film.js
@@ -1,4 +1,5 @@
 import dayjs from "dayjs";
+import {getRandomInteger, getRandomArrayItem, getArrayWithRandomItems} from "../utils.js";
 
 const MAX_COMMENT = 5;
 const MAX_RATING = 10;
@@ -41,25 +42,6 @@ const DESCRIPTION_TEXTS = [
   `Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus.`,
   `In rutrum ac purus sit amet tempus.`
 ];
-
-const getRandomInteger = (a = 0, b = 1) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-const getRandomArrayItem = (arr) => {
-  const randomIndex = getRandomInteger(0, arr.length - 1);
-
-  return arr[randomIndex];
-};
-
-const getArrayWithRandomItems = (arr) => {
-  return arr.filter(() => {
-    return getRandomInteger();
-  });
-};
 
 const generateDescription = () => {
   return getArrayWithRandomItems(DESCRIPTION_TEXTS).join(` `);
